% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_WEPPeval_application}
\alias{run_WEPPeval_application}
\title{WEPPeval shiny application to be run locally}
\usage{
run_WEPPeval_application()
}
\description{
WEPPeval shiny application to be run locally
}
